Creation test fix

+++ BackEnd/sept-booking/src/test/java/sept/rmit/booking/BookingApplicationTests.java
@@ -42,8 +42,9 @@ class BookingApplicationTests {
 
         assertEquals(mockBooking, newBooking);
 
-        repo.deleteById(newBooking.getId());
-        repo.deleteById(mockBooking.getId());
+        if(newBooking != null){
+            repo.deleteById(newBooking.getId());
+        }
     }
 
     @DisplayName("Booking Exist(True)")
